+++ sfeos_tools/cli_options.py
@@ -104,6 +104,19 @@ def stac_api_options(f):
     )(f)
     return f
 
+def lang_options(f):
+    """Add language option to CLI command.
+
+    Adds the following option:
+    - --lang: Language of raw data (default: None)
+    """
+    f = click.option(
+        "--lang",
+        type=str,
+        default=None,
+        help="Language of raw data (default: None)",
+    )(f)
+    return f
 
 def set_es_env_vars(
     host=None, port=None, use_ssl=None, user=None, password=None, api_key=None